chore(lint): make clippy happy
